# docs/openapi.yaml
openapi: 3.0.3
info:
  title: UAV Docking Management System API
  description: |
    Comprehensive REST API for managing UAV operations, real-time tracking, geofencing, and fleet management.
    
    ## Authentication
    The API uses role-based authentication with three user roles:
    - **USER**: Read-only access
    - **OPERATOR**: Can perform UAV operations and updates  
    - **ADMIN**: Full system access including delete operations
    
    ## Rate Limiting
    - Default: 100 requests per minute
    - Admin: 1000 requests per minute
    - API Key: 2000 requests per minute
  version: 1.0.0
  contact:
    name: UAV Management System Team
    email: support@uavmanagement.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.uavmanagement.com
    description: Production server
security:
  - sessionAuth: []
  - apiKeyAuth: []
paths:
...
  /api/docking-stations:
    get:
      tags:
        - Docking Stations
      summary: Get all docking stations
      description: Retrieve all docking stations
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Docking stations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DockingStation'
    post:
      tags:
        - Docking Stations
      summary: Create docking station
      description: Create a new docking station (Admin only)
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DockingStationCreateRequest'
      responses:
        '201':
          description: Docking station created successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    DockingStation:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Station Alpha"
        description:
          type: string
          example: "Primary docking station for Zone A"
        latitude:
          type: number
          format: double
          example: 40.7589
        longitude:
          type: number
          format: double
          example: -73.9851
        capacity:
          type: integer
          example: 4
        currentOccupancy:
          type: integer
          example: 2
        status:
          type: string
          enum: [OPERATIONAL, MAINTENANCE, OFFLINE, FULL]
          example: "OPERATIONAL"
        stationType:
          type: string
          enum: [CHARGING, MAINTENANCE, STORAGE, MULTI_PURPOSE]
          example: "CHARGING"
    DockingStationCreateRequest:
      type: object
      required:
        - name
        - latitude
        - longitude
        - capacity
      properties:
        name:
          type: string
          example: "Station Alpha"
        description:
          type: string
          example: "Primary docking station for Zone A"
        latitude:
          type: number
          format: double
          example: 40.7589
        longitude:
          type: number
          format: double
          example: -73.9851
        capacity:
          type: integer
          minimum: 1
          example: 4
        stationType:
          type: string
          enum: [CHARGING, MAINTENANCE, STORAGE, MULTI_PURPOSE]
          example: "CHARGING"
        operationalHours:
          type: string
          example: "24/7"
        contactInfo:
          type: string
          example: "station-alpha@company.com"
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "UAV_NOT_FOUND"
            message:
              type: string
              example: "UAV with specified ID not found"
            details:
              type: object
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
  responses:
    BadRequest:
      description: Bad request - invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...